+++ crates/teloxide-core/src/payloads.rs
@@ -53,6 +53,7 @@ mod edit_forum_topic;
 mod edit_general_forum_topic;
 mod edit_message_caption;
 mod edit_message_caption_inline;
+mod edit_message_checklist;
 mod edit_message_live_location;
 mod edit_message_live_location_inline;
 mod edit_message_media;
@@ -86,6 +87,7 @@ mod get_my_default_administrator_rights;
 mod get_my_description;
 mod get_my_name;
 mod get_my_short_description;
+mod get_my_star_balance;
 mod get_star_transactions;
 mod get_sticker_set;
 mod get_updates;
@@ -114,6 +116,7 @@ mod save_prepared_inline_message;
 mod send_animation;
 mod send_audio;
 mod send_chat_action;
+mod send_checklist;
 mod send_contact;
 mod send_dice;
 mod send_document;
@@ -221,6 +224,7 @@ pub use edit_forum_topic::{EditForumTopic, EditForumTopicSetters};
 pub use edit_general_forum_topic::{EditGeneralForumTopic, EditGeneralForumTopicSetters};
 pub use edit_message_caption::{EditMessageCaption, EditMessageCaptionSetters};
 pub use edit_message_caption_inline::{EditMessageCaptionInline, EditMessageCaptionInlineSetters};
+pub use edit_message_checklist::{EditMessageChecklist, EditMessageChecklistSetters};
 pub use edit_message_live_location::{EditMessageLiveLocation, EditMessageLiveLocationSetters};
 pub use edit_message_live_location_inline::{
     EditMessageLiveLocationInline, EditMessageLiveLocationInlineSetters,
@@ -264,6 +268,7 @@ pub use get_my_default_administrator_rights::{
 pub use get_my_description::{GetMyDescription, GetMyDescriptionSetters};
 pub use get_my_name::{GetMyName, GetMyNameSetters};
 pub use get_my_short_description::{GetMyShortDescription, GetMyShortDescriptionSetters};
+pub use get_my_star_balance::{GetMyStarBalance, GetMyStarBalanceSetters};
 pub use get_star_transactions::{GetStarTransactions, GetStarTransactionsSetters};
 pub use get_sticker_set::{GetStickerSet, GetStickerSetSetters};
 pub use get_updates::{GetUpdates, GetUpdatesSetters};
@@ -296,6 +301,7 @@ pub use save_prepared_inline_message::{
 pub use send_animation::{SendAnimation, SendAnimationSetters};
 pub use send_audio::{SendAudio, SendAudioSetters};
 pub use send_chat_action::{SendChatAction, SendChatActionSetters};
+pub use send_checklist::{SendChecklist, SendChecklistSetters};
 pub use send_contact::{SendContact, SendContactSetters};
 pub use send_dice::{SendDice, SendDiceSetters};
 pub use send_document::{SendDocument, SendDocumentSetters};