+++ crates/teloxide-core/src/types/message.rs
@@ -7,16 +7,16 @@ use url::Url;
 
 use crate::types::{
     Animation, Audio, BareChatId, BusinessConnectionId, Chat, ChatBackground, ChatBoostAdded,
-    ChatId, ChatShared, Contact, Dice, Document, ExternalReplyInfo, ForumTopicClosed,
-    ForumTopicCreated, ForumTopicEdited, ForumTopicReopened, Game, GeneralForumTopicHidden,
-    GeneralForumTopicUnhidden, GiftInfo, Giveaway, GiveawayCompleted, GiveawayCreated,
-    GiveawayWinners, InlineKeyboardMarkup, Invoice, LinkPreviewOptions, Location,
-    MaybeInaccessibleMessage, MessageAutoDeleteTimerChanged, MessageEntity, MessageEntityRef,
-    MessageId, MessageOrigin, PaidMediaInfo, PaidMessagePriceChanged, PassportData, PhotoSize,
-    Poll, ProximityAlertTriggered, RefundedPayment, Sticker, Story, SuccessfulPayment, TextQuote,
-    ThreadId, True, UniqueGiftInfo, User, UsersShared, Venue, Video, VideoChatEnded,
-    VideoChatParticipantsInvited, VideoChatScheduled, VideoChatStarted, VideoNote, Voice,
-    WebAppData, WriteAccessAllowed,
+    ChatId, ChatShared, Checklist, ChecklistTasksAdded, ChecklistTasksDone, Contact, Dice,
+    DirectMessagePriceChanged, Document, ExternalReplyInfo, ForumTopicClosed, ForumTopicCreated,
+    ForumTopicEdited, ForumTopicReopened, Game, GeneralForumTopicHidden, GeneralForumTopicUnhidden,
+    GiftInfo, Giveaway, GiveawayCompleted, GiveawayCreated, GiveawayWinners, InlineKeyboardMarkup,
+    Invoice, LinkPreviewOptions, Location, MaybeInaccessibleMessage, MessageAutoDeleteTimerChanged,
+    MessageEntity, MessageEntityRef, MessageId, MessageOrigin, PaidMediaInfo,
+    PaidMessagePriceChanged, PassportData, PhotoSize, Poll, ProximityAlertTriggered,
+    RefundedPayment, Sticker, Story, SuccessfulPayment, TextQuote, ThreadId, True, UniqueGiftInfo,
+    User, UsersShared, Venue, Video, VideoChatEnded, VideoChatParticipantsInvited,
+    VideoChatScheduled, VideoChatStarted, VideoNote, Voice, WebAppData, WriteAccessAllowed,
 };
 
 /// This object represents a message.
@@ -95,6 +95,9 @@ pub enum MessageKind {
     ProximityAlertTriggered(MessageProximityAlertTriggered),
     ChatBoostAdded(MessageChatBoostAdded),
     ChatBackground(MessageChatBackground),
+    ChecklistTasksDone(MessageChecklistTasksDone),
+    ChecklistTasksAdded(MessageChecklistTasksAdded),
+    DirectMessagePriceChanged(MessageDirectMessagePriceChanged),
     ForumTopicCreated(MessageForumTopicCreated),
     ForumTopicEdited(MessageForumTopicEdited),
     ForumTopicClosed(MessageForumTopicClosed),
@@ -397,6 +400,7 @@ pub enum MediaKind {
     Location(MediaLocation),
     Photo(MediaPhoto),
     Poll(MediaPoll),
+    Checklist(MediaChecklist),
     Sticker(MediaSticker),
     Story(MediaStory),
     Text(MediaText),
@@ -553,6 +557,13 @@ pub struct MediaPoll {
 
 #[serde_with::skip_serializing_none]
 #[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
+pub struct MediaChecklist {
+    /// Message is a checklist, information about the checklist.
+    pub checklist: Checklist,
+}
+
+#[serde_with::skip_serializing_none]
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 pub struct MediaSticker {
     /// Message is a sticker, information about the sticker.
     pub sticker: Sticker,
@@ -667,6 +678,29 @@ pub struct MessageChatBoostAdded {
 pub struct MessageChatBackground {
     /// Service message. Chat background set.
     pub chat_background_set: ChatBackground,
+}
+
+#[serde_with::skip_serializing_none]
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
+pub struct MessageChecklistTasksDone {
+    /// Service message: some tasks in a checklist were marked as done or not
+    /// done.
+    pub checklist_tasks_done: ChecklistTasksDone,
+}
+
+#[serde_with::skip_serializing_none]
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
+pub struct MessageChecklistTasksAdded {
+    /// Service message: tasks were added to a checklist.
+    pub checklist_tasks_added: ChecklistTasksAdded,
+}
+
+#[serde_with::skip_serializing_none]
+#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
+pub struct MessageDirectMessagePriceChanged {
+    /// Service message: the price for paid messages in the corresponding direct
+    /// messages chat of a channel has changed.
+    pub direct_message_price_changed: DirectMessagePriceChanged,
 }
 
 #[serde_with::skip_serializing_none]
@@ -821,15 +855,17 @@ mod getters {
 
     use crate::types::{
         self, message::MessageKind::*, Chat, ChatId, ChatMigration, EffectId, LinkPreviewOptions,
-        MaybeInaccessibleMessage, MediaAnimation, MediaAudio, MediaContact, MediaDocument,
-        MediaGame, MediaKind, MediaLocation, MediaPaid, MediaPhoto, MediaPoll, MediaSticker,
-        MediaStory, MediaText, MediaVenue, MediaVideo, MediaVideoNote, MediaVoice, Message,
-        MessageChannelChatCreated, MessageChatShared, MessageCommon, MessageConnectedWebsite,
-        MessageDeleteChatPhoto, MessageDice, MessageEntity, MessageGroupChatCreated, MessageId,
-        MessageInvoice, MessageLeftChatMember, MessageNewChatMembers, MessageNewChatPhoto,
-        MessageNewChatTitle, MessageOrigin, MessagePassportData, MessagePinned,
-        MessageProximityAlertTriggered, MessageSuccessfulPayment, MessageSupergroupChatCreated,
-        MessageUsersShared, MessageVideoChatParticipantsInvited, PhotoSize, Story, TextQuote, User,
+        MaybeInaccessibleMessage, MediaAnimation, MediaAudio, MediaChecklist, MediaContact,
+        MediaDocument, MediaGame, MediaKind, MediaLocation, MediaPaid, MediaPhoto, MediaPoll,
+        MediaSticker, MediaStory, MediaText, MediaVenue, MediaVideo, MediaVideoNote, MediaVoice,
+        Message, MessageChannelChatCreated, MessageChatShared, MessageChecklistTasksAdded,
+        MessageChecklistTasksDone, MessageCommon, MessageConnectedWebsite, MessageDeleteChatPhoto,
+        MessageDice, MessageDirectMessagePriceChanged, MessageEntity, MessageGroupChatCreated,
+        MessageId, MessageInvoice, MessageLeftChatMember, MessageNewChatMembers,
+        MessageNewChatPhoto, MessageNewChatTitle, MessageOrigin, MessagePassportData,
+        MessagePinned, MessageProximityAlertTriggered, MessageSuccessfulPayment,
+        MessageSupergroupChatCreated, MessageUsersShared, MessageVideoChatParticipantsInvited,
+        PhotoSize, Story, TextQuote, User,
     };
 
     use super::{
@@ -1106,6 +1142,7 @@ mod getters {
                     | MediaKind::Venue(_)
                     | MediaKind::Location(_)
                     | MediaKind::Poll(_)
+                    | MediaKind::Checklist(_)
                     | MediaKind::Sticker(_)
                     | MediaKind::Story(_)
                     | MediaKind::Text(_)
@@ -1137,6 +1174,7 @@ mod getters {
                     | MediaKind::Venue(_)
                     | MediaKind::Location(_)
                     | MediaKind::Poll(_)
+                    | MediaKind::Checklist(_)
                     | MediaKind::Sticker(_)
                     | MediaKind::Story(_)
                     | MediaKind::Text(_)
@@ -1325,6 +1363,17 @@ mod getters {
                     media_kind: MediaKind::Poll(MediaPoll { poll, .. }),
                     ..
                 }) => Some(poll),
+                _ => None,
+            }
+        }
+
+        #[must_use]
+        pub fn checklist(&self) -> Option<&types::Checklist> {
+            match &self.kind {
+                Common(MessageCommon {
+                    media_kind: MediaKind::Checklist(MediaChecklist { checklist, .. }),
+                    ..
+                }) => Some(checklist),
                 _ => None,
             }
         }
@@ -1587,6 +1636,36 @@ mod getters {
                 ChatBackground(MessageChatBackground { chat_background_set }) => {
                     Some(chat_background_set)
                 }
+                _ => None,
+            }
+        }
+
+        #[must_use]
+        pub fn checklist_tasks_done(&self) -> Option<&types::ChecklistTasksDone> {
+            match &self.kind {
+                ChecklistTasksDone(MessageChecklistTasksDone { checklist_tasks_done }) => {
+                    Some(checklist_tasks_done)
+                }
+                _ => None,
+            }
+        }
+
+        #[must_use]
+        pub fn checklist_tasks_added(&self) -> Option<&types::ChecklistTasksAdded> {
+            match &self.kind {
+                ChecklistTasksAdded(MessageChecklistTasksAdded { checklist_tasks_added }) => {
+                    Some(checklist_tasks_added)
+                }
+                _ => None,
+            }
+        }
+
+        #[must_use]
+        pub fn direct_message_price_changed(&self) -> Option<&types::DirectMessagePriceChanged> {
+            match &self.kind {
+                DirectMessagePriceChanged(MessageDirectMessagePriceChanged {
+                    direct_message_price_changed,
+                }) => Some(direct_message_price_changed),
                 _ => None,
             }
         }